+++ examples/Old Format/mv_setup_options.py
@@ -1,11 +1,9 @@
 from __future__ import print_function
 from sympy import symbols
 from galgebra.deprecated import MV
-from galgebra.printer import enhance_print,Get_Program,Print_Function
+from galgebra.printer import enhance_print,Print_Function
 
 def MV_setup_options():
-    Print_Function()
-
     (e1,e2,e3) = MV.setup('e_1 e_2 e_3','[1,1,1]')
     v = MV('v', 'vector')
     print(v)
@@ -26,7 +24,6 @@ def MV_setup_options():
     return
 
 def main():
-    Get_Program(True)
     enhance_print()
     MV_setup_options()
     return